+++ changelog.md
@@ -10,7 +10,7 @@
       - Add ctrl+click selection system
       - Fix code generation of room sounds and smells
       - Add hover previews to display monster look/item longs for room preview and monster preview panels
-      - Add includes to area properties
+      - Add includes and defines to area properties
       - Add functions to rooms
       - Add standard object support to add money, gems, and standard objects
       - Add argument support to objects and monsters to allow passing arguments when object is created